+++ iceoryx_posh/source/popo/ports/publisher_port_roudi.cpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2020 by Robert Bosch GmbH. All rights reserved.
-// Copyright (c) 2021 by Apex.AI Inc. All rights reserved.
+// Copyright (c) 2021 - 2022 by Apex.AI Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -57,10 +57,7 @@ cxx::optional<capro::CaproMessage> PublisherPortRouDi::tryGetCaProMessage() noex
 
         const auto historyCapacity = m_chunkSender.getHistoryCapacity();
         caproMessage.m_historyCapacity = historyCapacity;
-
-        // provide additional AUTOSAR Adaptive like information
-        caproMessage.m_subType =
-            (0u < historyCapacity) ? capro::CaproMessageSubType::FIELD : capro::CaproMessageSubType::EVENT;
+        caproMessage.m_serviceType = capro::CaproServiceType::PUBLISHER;
 
         return cxx::make_optional<capro::CaproMessage>(caproMessage);
     }
@@ -72,6 +69,8 @@ cxx::optional<capro::CaproMessage> PublisherPortRouDi::tryGetCaProMessage() noex
         m_chunkSender.removeAllQueues();
 
         capro::CaproMessage caproMessage(capro::CaproMessageType::STOP_OFFER, this->getCaProServiceDescription());
+        caproMessage.m_serviceType = capro::CaproServiceType::PUBLISHER;
+
         return cxx::make_optional<capro::CaproMessage>(caproMessage);
     }
     else
@@ -85,7 +84,7 @@ cxx::optional<capro::CaproMessage>
 PublisherPortRouDi::dispatchCaProMessageAndGetPossibleResponse(const capro::CaproMessage& caProMessage) noexcept
 {
     capro::CaproMessage responseMessage(
-        capro::CaproMessageType::NACK, this->getCaProServiceDescription(), capro::CaproMessageSubType::NOSUBTYPE);
+        capro::CaproMessageType::NACK, this->getCaProServiceDescription(), capro::CaproServiceType::NONE);
 
     if (getMembers()->m_offered.load(std::memory_order_relaxed))
     {